pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "amon-hen"
version = "0.1.0"
description = "Narrative intelligence platform — The Seat of Seeing"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [{ name = "civilian24601" }]
dependencies = [
    # Core
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "httpx>=0.27",
    "apscheduler>=3.10,<4.0",
    "click>=8.0",
    "python-dotenv>=1.0",
    "pyyaml>=6.0",

    # LLM
    "anthropic>=0.40",

    # Embeddings & Vector
    "sentence-transformers>=5.0",
    "qdrant-client>=1.12",

    # Clustering & ML
    "scikit-learn>=1.5",
    "numpy>=1.26",

    # Sources
    "feedparser>=6.0",
    "atproto>=0.0.55",
    "praw>=7.7",
    "gdeltdoc>=1.10",

    # Web
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "jinja2>=3.1",
]

[project.optional-dependencies]
viz = [
    "umap-learn>=0.5",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
    "ruff>=0.8",
    "mypy>=1.13",
]

[project.scripts]
amon = "amon_hen.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/amon_hen"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.11"
strict = true
